feat: use dynamodb instead of ssm for JIT config
Add extra capabilities to support ephemeral multiarch runners.
In workflow jobs with large matrix, github can request large number of
runners simultaneously. This can cause the ssm to be overwhelmed.
This PR replaces the ssm with dynamodb to store the JIT config.

Also fixed a couple things:
- `runners_ssm_housekeeper` schema was not aligned across modules
- syntax for `env` in `github_agent.ubuntu.pkr.hcl`

Added a couple of improvements:
- updated baseline provisioning script for prebuilt runners to make it
closer to what github hosted runners do
- startup-runner.sh now automatically discovers attached ebs volume
and mounts it to /data