feat: add lambda function to cleanup org runners
Add a lambda function to cleanup offline runners in a GitHub
organization. Normally runners will be cleaned up automatically, but
when using ephemeral runners and spot instances, the call to Github
API to remove the runner may not happen, and the runner will stay in
the list in offline state.
This lambda function will be triggered by a CloudWatch event and
will remove any organization runners that are offline, and their
labels match the config.