fix: Tag non jit instances with runner id and add docs (#4667)
⚠️ This PR is on top of #4539

This pull request is an enhancement on top of #4539. The PR also
introduces tagging via the start scripts for non-JIT configured
instances. And it adds a state diagram to the ever growing complexity of
the state diagram below terminating instances.

### Instance Tagging Enhancements:
* Add function to tag runners with runner id for non JIT instances.
Functions are designed to ignore errors to avoid causing a crash of the
runner registration process.
* side effect is that the instance is allowed to create the tag
`ghr:github_runner_id` the instance is allowed to create the tag only on
itself.

### Added docs
* State diagram for scale-down added

<img width="2084" height="3840" alt="Untitled diagram _ Mermaid
Chart-2025-07-20-140732"
src="https://github.com/user-attachments/assets/e88af647-98e5-46cb-8da4-28110c608d8d"
/>

---------

Co-authored-by: github-aws-runners-pr|bot <github-aws-runners-pr[bot]@users.noreply.github.com>