style: minor tweaks to `ScaleError`
Some small tweaks to the style of `ScaleError`, while we're working on
it.

`message` doesn't need to be a public property. `Error` already has it.
Likewise with `stack` - we can use the parent's implementation.

Instead of `getDetailedMessage` we can have a property getter which is
very slightly nicer to read at callsites (`e.detailedMessage` vs.
`e.getDetailedMessage()`).