fix: only retry when we don't create enough ephemeral runners
Prior to this change, we would have retried if _any_ runner failed to be
created either due to capacity or an EC2 error. Now we'll consider these
differently.

If we can't create due to the maximum allowed runner count, retry only
if creating ephemeral runners.

Always retry if EC2 creation fails.