feat(scale-down): use SSM to store scale-down config
Now we're potentially running multiple configurations in one scale-down
invocation, if we continue to use the environment we could start to hit
size limits: on Lambda, environment variables are limited to 4K.

Adopt the approach we use elsewhere and switch to SSM parameter store
for config. Here we add all the necessary IAM permissions, arrange to
store the config in the store and then read it back in `scale-down`.

A more strict parser is also introduced, ensuring that we detect more
invalid configurations and reject them with clear error messages.